Pull out box related functionality into a separate module
