+++ src/App.js
@@ -21,9 +21,8 @@ import Config from "./Config";
 import ScrollToTop from "./components/ScrollToTop";
 import ErrorBoundary from "./components/ErrorBoundary";
 import DashboardPage from "./pages/DashboardPage";
-import BoxesPage from "./pages/BoxesPage";
+import { BoxesPage, CreateLabelsPage } from "./modules/boxes";
 import HomePage from "./pages/HomePage";
-import CreateLabelsPage from "./pages/CreateLabelsPage";
 import { ProductsPage } from "./modules/products";
 import CreateOrganizationPage from "./pages/CreateOrganizationPage";
 import InvitePage from "./pages/InvitePage";
