+++ src/modules/boxes/components/AddBoxDialog.js
@@ -2,7 +2,7 @@ import React, { PureComponent } from "react";
 import { bool, func } from "prop-types";
 import Dialog from "@material-ui/core/Dialog";
 
-import { ProductsCollection } from "queries/products";
+import { ProductsCollection } from "modules/products/queries";
 
 import AddBoxForm from "./AddBoxForm";
 import AddBoxDone from "./AddBoxDone";
