+++ src/modules/boxes/index.js
@@ -1,0 +1,3 @@
+export { default as CreateLabelsPage } from "./pages/CreateLabelsPage";
+export { default as BoxesPage } from "./pages/BoxesPage";
+export { default as AddBoxDialog } from "./components/AddBoxDialog";
