+++ src/modules/boxes/pages/BoxesPage.js
@@ -1,7 +1,8 @@
 import React from "react";
 
 import AppFrame from "components/AppFrame";
-import BoxListContainer from "components/BoxListContainer";
+
+import BoxListContainer from "../components/BoxListContainer";
 
 class BoxesPage extends React.Component {
   render() {
