+++ src/reducers/index.js
@@ -2,9 +2,9 @@ import { combineReducers } from "redux";
 
 import user from "modules/auth/reducer";
 import products from "modules/products/reducer";
+import boxes from "modules/boxes/reducer";
 
 import profile from "./profile";
-import boxes from "./boxes";
 
 export default combineReducers({
   products,
